% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_country_class}
\alias{add_country_class}
\title{Add a \code{country_class} column based on the \code{Country} column}
\usage{
add_country_class(data, oecd_countries)
}
\arguments{
\item{data}{Data frame with a \code{Country} column.}

\item{oecd_countries}{Data frame with a \code{Name} column of OECD countries.}
}
\value{
\code{data} with a \code{country_class} column (see \code{\link[=classify_country]{classify_country()}}).
}
\description{
Add a \code{country_class} column based on the \code{Country} column
}
