% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{classify_country}
\alias{classify_country}
\title{Classify countries into UK, other OECD and rest of world}
\usage{
classify_country(country, oecd_names)
}
\arguments{
\item{country}{Character vector of country names.}

\item{oecd_names}{Character vector of OECD country names.}
}
\value{
Character vector with values \code{"United Kingdom"}, \code{"Other OECD"}
or \code{"Rest of World"}.
}
\description{
Classify countries into UK, other OECD and rest of world
}
