% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_retraction_rates}
\alias{plot_retraction_rates}
\title{Plot retraction rates per year by country class and scenario}
\usage{
plot_retraction_rates(retraction_rates, year_limits = c(2000, 2023))
}
\arguments{
\item{retraction_rates}{Output of \code{\link[=compute_retraction_rates]{compute_retraction_rates()}}.}

\item{year_limits}{Range of publication years shown.}
}
\value{
A ggplot object.
}
\description{
Plot retraction rates per year by country class and scenario
}
