% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_significant_reasons}
\alias{plot_significant_reasons}
\title{Dumbbell plot of the proportions of significant reasons in two groups}
\usage{
plot_significant_reasons(sig_long)
}
\arguments{
\item{sig_long}{Output of \code{\link[=significant_reasons_long]{significant_reasons_long()}}.}
}
\value{
A ggplot object.
}
\description{
Dumbbell plot of the proportions of significant reasons in two groups
}
