% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractionwatch-package.R
\docType{package}
\name{retractionwatch-package}
\alias{retractionwatch}
\alias{retractionwatch-package}
\title{retractionwatch: Analysis of Retraction Watch Data by Country Group}
\description{
Functions to load, clean and analyse the Retraction Watch dataset together with SCImago country publication counts. Includes retraction rates by country group under several filtering scenarios, comparisons of retraction reasons between groups, and analyses of AI-related retractions.
}
\author{
\strong{Maintainer}: Miquel Perello Nieto \email{mp15688@bristol.ac.uk}

Authors:
\itemize{
  \item Miquel Perello Nieto \email{mp15688@bristol.ac.uk}
}

}
\keyword{internal}
